refactor(cache): split NotificationListenerService into partial classes (#432)
Pure file split, no behavior change. Mirrors the partial-class recipe
used in PR #428 (CirclesRpcModule), #429 (CacheWarmupService), #431
(SimulationCanaryService).

NotificationListenerService.cs: 1301 → 366 LoC (shell, lifecycle,
handler, recovery, helpers, top-level dispatch).

New per-table partials under Services/Listeners/:
- NotificationListenerService.V1Events.cs   — V1 signups + transfers + trust + metadata
- NotificationListenerService.V2Registry.cs — V2 dispatcher + RegisterHuman/Organization/Group
- NotificationListenerService.V2Trust.cs    — V2 Trust + GroupMemberships
- NotificationListenerService.V2Wrappers.cs — ERC20WrapperDeployed
- NotificationListenerService.V2Transfers.cs — TransferSingle/Batch + ERC20 wrapper transfers
- NotificationListenerService.V2Metadata.cs — UpdateMetadataDigest + RegisterShortName + SetAdvancedUsageFlag

Test coupling bundled:
- PathfinderGraphControllerTests.NotificationListener_V2TransferSql_UsesTokenAddressNotId
  now searches all NotificationListenerService*.cs partials recursively
  for the method definition. The test asserts source-code properties of
  ProcessV2TransfersAsync which moved to V2Transfers.cs.

Verification (clean rebuild against staging baseline):
- Build: 0 errors / 60 warnings (matches baseline)
- Tests: 226 passed / 0 failed / 26 skipped / 252 total (matches baseline)
- Public method surface diff: empty
- Source-level method body equivalence (whitespace-normalized SHA256):
  24/24 identical across baseline and post-split
