+++ src/Cache/Circles.Cache.Service/Services/NotificationListenerService.cs
@@ -1,4 +1,3 @@
-using System.Numerics;
 using Circles.Cache.Service.Caches;
 using Circles.Cache.Service.Metrics;
 using Circles.Common;
@@ -10,8 +9,17 @@ namespace Circles.Cache.Service.Services;
 /// Background service that listens to PostgreSQL NOTIFY events from the Indexer.
 /// Treats notifications as "pings" and queries the database directly for block information.
 /// Uses BlockRingBuffer to detect chain reorganizations.
+///
+/// The implementation is split across multiple partial-class files for maintainability:
+/// - NotificationListenerService.cs                       — fields, ctor, lifecycle (ExecuteAsync, ListenForNotificationsAsync, HandleNotificationAsync), recovery (AttemptRecoveryRollbackAsync, ClearAllCaches, TriggerFullRewarmup), shared helpers (GetBlockTimestampAsync, WithReadonlyConnectionAsync, GetRecentBlocksAsync), top-level dispatch (ProcessBlockRangeAsync)
+/// - Listeners/NotificationListenerService.V1Events.cs    — ProcessV1EventsAsync (incl. inline signups), ProcessV1TransfersAsync, ProcessV1TrustAsync, ProcessV1UpdateMetadataDigestAsync
+/// - Listeners/NotificationListenerService.V2Registry.cs  — ProcessV2EventsAsync, ProcessV2RegisterHumanAsync, ProcessV2RegisterOrganizationAsync, ProcessV2RegisterGroupAsync
+/// - Listeners/NotificationListenerService.V2Trust.cs     — ProcessV2TrustAsync
+/// - Listeners/NotificationListenerService.V2Wrappers.cs  — ProcessV2Erc20WrapperDeployedAsync
+/// - Listeners/NotificationListenerService.V2Transfers.cs — ProcessV2TransfersAsync, ProcessV2Erc20WrapperTransfersAsync
+/// - Listeners/NotificationListenerService.V2Metadata.cs  — ProcessV2UpdateMetadataDigestAsync, ProcessV2RegisterShortNameAsync, ProcessV2SetAdvancedUsageFlagAsync
 /// </summary>
-public class NotificationListenerService : BackgroundService
+public partial class NotificationListenerService : BackgroundService
 {
     private readonly ILogger<NotificationListenerService> _logger;
     private readonly CacheServiceSettings _settings;
@@ -424,878 +432,5 @@ public class NotificationListenerService : BackgroundService
 
         // Process V2 events in this range
         await ProcessV2EventsAsync(conn, fromBlock, toBlock, ct);
-    }
-
-    private async Task ProcessV1EventsAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        // Process V1 Signups
-        const string humanSignupSql = @"
-            SELECT s.""blockNumber"", s.""user"", s.""token""
-            FROM ""CrcV1_Signup"" s
-            WHERE s.""blockNumber"" >= @fromBlock AND s.""blockNumber"" <= @toBlock
-            ORDER BY s.""blockNumber"", s.""transactionIndex"", s.""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(humanSignupSql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var count = 0;
-
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var blockNumber = reader.GetInt64(0);
-                var user = reader.GetString(1);
-                var token = reader.GetString(2);
-
-                var userKey = user.ToLowerInvariant();
-                var tokenKey = token.ToLowerInvariant();
-
-                _caches.V1Avatars.Add(blockNumber, userKey, ("CrcV1_Signup", token));
-                _caches.V1TokenOwnerByToken.Add(blockNumber, tokenKey, user);
-                count++;
-            }
-        }
-
-        if (count > 0)
-        {
-            _logger.LogDebug("Processed {Count} V1 human signups", count);
-        }
-
-        // Process V1 Organization Signups
-        const string orgSignupSql = @"
-            SELECT o.""blockNumber"", o.""organization""
-            FROM ""CrcV1_OrganizationSignup"" o
-            WHERE o.""blockNumber"" >= @fromBlock AND o.""blockNumber"" <= @toBlock
-            ORDER BY o.""blockNumber"", o.""transactionIndex"", o.""logIndex""";
-
-        await using var orgCmd = new NpgsqlCommand(orgSignupSql, conn);
-        orgCmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        orgCmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var orgCount = 0;
-
-        await using (var orgReader = await orgCmd.ExecuteReaderAsync(ct))
-        {
-            while (await orgReader.ReadAsync(ct))
-            {
-                var blockNumber = orgReader.GetInt64(0);
-                var organization = orgReader.GetString(1);
-
-                var orgKey = organization.ToLowerInvariant();
-
-                _caches.V1Avatars.Add(blockNumber, orgKey, ("CrcV1_OrganizationSignup", null));
-                orgCount++;
-            }
-        }
-
-        if (orgCount > 0)
-        {
-            _logger.LogDebug("Processed {Count} V1 organization signups", orgCount);
-        }
-
-        // Process V1 Transfers (for balance updates)
-        await ProcessV1TransfersAsync(conn, fromBlock, toBlock, ct);
-
-        // Process V1 Trust events
-        await ProcessV1TrustAsync(conn, fromBlock, toBlock, ct);
-
-        // Process V1 UpdateMetadataDigest (for CID maps)
-        await ProcessV1UpdateMetadataDigestAsync(conn, fromBlock, toBlock, ct);
-    }
-
-    private async Task ProcessV2EventsAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        // Process V2 RegisterHuman
-        await ProcessV2RegisterHumanAsync(conn, fromBlock, toBlock, ct);
-
-        // Process V2 RegisterOrganization
-        await ProcessV2RegisterOrganizationAsync(conn, fromBlock, toBlock, ct);
-
-        // Process V2 RegisterGroup
-        await ProcessV2RegisterGroupAsync(conn, fromBlock, toBlock, ct);
-
-        // Note: CrcV2_Stopped is NOT processed here — stop() only prevents minting,
-        // it does not deregister the avatar. Stopped avatars remain in V2Avatars/Groups.
-
-        // Process V2 Trust (for group memberships)
-        await ProcessV2TrustAsync(conn, fromBlock, toBlock, ct);
-
-        // Process V2 ERC20WrapperDeployed
-        await ProcessV2Erc20WrapperDeployedAsync(conn, fromBlock, toBlock, ct);
-
-        // Process V2 Transfers (for balance updates)
-        await ProcessV2TransfersAsync(conn, fromBlock, toBlock, ct);
-
-        // Process V2 ERC20 Wrapper Transfers (for balance updates)
-        await ProcessV2Erc20WrapperTransfersAsync(conn, fromBlock, toBlock, ct);
-
-        // Process V2 UpdateMetadataDigest (for CID maps)
-        await ProcessV2UpdateMetadataDigestAsync(conn, fromBlock, toBlock, ct);
-
-        // Process V2 RegisterShortName (for short name mappings)
-        await ProcessV2RegisterShortNameAsync(conn, fromBlock, toBlock, ct);
-
-        // Process V2 SetAdvancedUsageFlag (for consented flow)
-        await ProcessV2SetAdvancedUsageFlagAsync(conn, fromBlock, toBlock, ct);
-    }
-
-    private async Task ProcessV2RegisterHumanAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        const string sql = @"
-            SELECT r.""blockNumber"", r.""timestamp"", r.""avatar""
-            FROM ""CrcV2_RegisterHuman"" r
-            WHERE r.""blockNumber"" >= @fromBlock AND r.""blockNumber"" <= @toBlock
-            ORDER BY r.""blockNumber"", r.""transactionIndex"", r.""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var count = 0;
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var blockNumber = reader.GetInt64(0);
-                var timestamp = reader.GetInt64(1);
-                var avatar = reader.GetString(2);
-
-                var avatarKey = avatar.ToLowerInvariant();
-
-                _caches.V2Avatars.Add(blockNumber, avatarKey, ("CrcV2_RegisterHuman", timestamp));
-                count++;
-            }
-        }
-
-        if (count > 0)
-        {
-            _logger.LogDebug("Processed {Count} V2 human registrations", count);
-        }
-    }
-
-    private async Task ProcessV2RegisterOrganizationAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        const string sql = @"
-            SELECT r.""blockNumber"", r.""timestamp"", r.""organization""
-            FROM ""CrcV2_RegisterOrganization"" r
-            WHERE r.""blockNumber"" >= @fromBlock AND r.""blockNumber"" <= @toBlock
-            ORDER BY r.""blockNumber"", r.""transactionIndex"", r.""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var count = 0;
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var blockNumber = reader.GetInt64(0);
-                var timestamp = reader.GetInt64(1);
-                var organization = reader.GetString(2);
-
-                var orgKey = organization.ToLowerInvariant();
-
-                _caches.V2Avatars.Add(blockNumber, orgKey, ("CrcV2_RegisterOrganization", timestamp));
-                count++;
-            }
-        }
-
-        if (count > 0)
-        {
-            _logger.LogDebug("Processed {Count} V2 organization registrations", count);
-        }
-    }
-
-    private async Task ProcessV2RegisterGroupAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        const string sql = @"
-            SELECT r.""blockNumber"", r.""group"", r.""name"", r.""mint"", r.""symbol""
-            FROM ""CrcV2_RegisterGroup"" r
-            WHERE r.""blockNumber"" >= @fromBlock AND r.""blockNumber"" <= @toBlock
-            ORDER BY r.""blockNumber"", r.""transactionIndex"", r.""logIndex""";
-
-        var count = 0;
-        await using (var cmd = new NpgsqlCommand(sql, conn))
-        {
-            cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-            cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-            await using (var groupReader = await cmd.ExecuteReaderAsync(ct))
-            {
-                while (await groupReader.ReadAsync(ct))
-                {
-                    var blockNumber = groupReader.GetInt64(0);
-                    var group = groupReader.GetString(1);
-                    var name = groupReader.GetString(2);
-                    var mint = groupReader.GetString(3);
-                    var symbol = groupReader.GetString(4);
-
-                    var groupKey = group.ToLowerInvariant();
-
-                    _caches.Groups.Add(blockNumber, groupKey, (name, mint, symbol));
-                    count++;
-                }
-            }
-        }
-
-        if (count > 0)
-        {
-            _logger.LogDebug("Processed {Count} V2 group registrations", count);
-        }
-    }
-
-    private async Task ProcessV2TrustAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        // Process all V2 Trust events - update both V2TrustRelations cache and GroupMemberships (when truster is a group)
-        const string sql = @"
-            SELECT
-                t.""blockNumber"",
-                t.""truster"",
-                t.""trustee"",
-                t.""expiryTime""
-            FROM ""CrcV2_Trust"" t
-            WHERE t.""blockNumber"" >= @fromBlock AND t.""blockNumber"" <= @toBlock
-            ORDER BY t.""blockNumber"", t.""transactionIndex"", t.""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var trustCount = 0;
-        var membershipCount = 0;
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var blockNumber = reader.GetInt64(0);
-                var truster = reader.GetString(1);
-                var trustee = reader.GetString(2);
-                var expiryTimeBig = reader.GetFieldValue<BigInteger>(3);
-
-                var trusterKey = truster.ToLowerInvariant();
-                var trusteeKey = trustee.ToLowerInvariant();
-
-                // Safely cast expiryTime to long
-                long expiryLong = expiryTimeBig > long.MaxValue ? long.MaxValue : (long)expiryTimeBig;
-
-                // Registration check: both truster and trustee must be registered avatars.
-                // Removals (expiryTime == 0) always proceed — safe to remove non-existent entries.
-                var trusterRegistered = _registrations.IsRegistered(trusterKey);
-                var trusteeRegistered = _registrations.IsRegistered(trusteeKey);
-
-                if (expiryTimeBig == 0)
-                {
-                    _caches.RemoveV2Trust(blockNumber, trusterKey, trusteeKey);
-                }
-                else if (trusterRegistered && trusteeRegistered)
-                {
-                    _caches.UpsertV2Trust(blockNumber, trusterKey, trusteeKey, expiryLong);
-                }
-                trustCount++;
-
-                // Also update GroupMemberships if truster is a group
-                if (_caches.Groups.ContainsKey(trusterKey))
-                {
-                    if (expiryTimeBig == 0)
-                    {
-                        _caches.RemoveGroupMembership(blockNumber, trusterKey, trusteeKey);
-                    }
-                    else if (trusteeRegistered)
-                    {
-                        _caches.UpsertGroupMembership(blockNumber, trusterKey, trusteeKey, expiryLong);
-                    }
-                    membershipCount++;
-                }
-            }
-        }
-
-        if (trustCount > 0)
-        {
-            _logger.LogDebug("Processed {TrustCount} V2 trust events ({MembershipCount} group memberships)",
-                trustCount, membershipCount);
-        }
-    }
-
-    private async Task ProcessV2Erc20WrapperDeployedAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        const string wrapperSql = @"
-            SELECT e.""blockNumber"", e.""avatar"", e.""erc20Wrapper"", e.""circlesType""
-            FROM ""CrcV2_ERC20WrapperDeployed"" e
-            WHERE e.""blockNumber"" >= @fromBlock AND e.""blockNumber"" <= @toBlock
-            ORDER BY e.""blockNumber"", e.""transactionIndex"", e.""logIndex""";
-
-        await using (var wrapperCmd = new NpgsqlCommand(wrapperSql, conn))
-        {
-            wrapperCmd.Parameters.AddWithValue("fromBlock", fromBlock);
-            wrapperCmd.Parameters.AddWithValue("toBlock", toBlock);
-
-            var count = 0;
-            await using (var wrapperReader = await wrapperCmd.ExecuteReaderAsync(ct))
-            {
-                while (await wrapperReader.ReadAsync(ct))
-                {
-                    var blockNumber = wrapperReader.GetInt64(0);
-                    var avatar = wrapperReader.GetString(1);
-                    var erc20Wrapper = wrapperReader.GetString(2);
-                    var circlesType = (CirclesType)wrapperReader.GetInt32(3);
-
-                    // Key by wrapper address (not avatar) to support avatars with multiple wrappers
-                    var wrapperKey = erc20Wrapper.ToLowerInvariant();
-                    var avatarKey = avatar.ToLowerInvariant();
-
-                    // Registration check: underlying avatar must be registered
-                    if (_registrations.IsRegistered(avatarKey))
-                    {
-                        _caches.UpsertWrapper(blockNumber, wrapperKey, avatar, circlesType);
-                        count++;
-                    }
-                }
-            }
-
-            if (count > 0)
-            {
-                _logger.LogDebug("Processed {Count} V2 ERC20 wrapper deployments", count);
-            }
-        }
-    }
-
-    private async Task ProcessV1TransfersAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        // Process V1 transfers incrementally
-        const string sql = @"
-            SELECT ""from"", ""to"", ""tokenAddress"", amount, ""blockNumber""
-            FROM ""CrcV1_Transfer""
-            WHERE ""blockNumber"" >= @fromBlock AND ""blockNumber"" <= @toBlock
-            ORDER BY ""blockNumber"", ""transactionIndex"", ""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var transferCount = 0;
-        var currentBalances = new Dictionary<string, decimal>();
-        long currentBlock = -1;
-
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var from = reader.GetString(0);
-                var to = reader.GetString(1);
-                var tokenAddress = reader.GetString(2);
-                var amountBig = reader.GetFieldValue<BigInteger>(3);
-                var blockNumber = reader.GetInt64(4);
-
-                var tokenKey = tokenAddress.ToLowerInvariant();
-                // Convert from wei (18 decimals) to token units using CirclesConverter for proper precision
-                decimal value;
-                try
-                {
-                    value = CirclesConverter.AttoCirclesToCircles(amountBig);
-                }
-                catch (OverflowException)
-                {
-                    _logger.LogWarning("Skipping V1 transfer with amount {Amount} that would overflow decimal", amountBig);
-                    continue;
-                }
-
-                if (blockNumber != currentBlock)
-                {
-                    if (currentBlock != -1)
-                    {
-                        // Flush balances for the previous block (index-before-balance ordering)
-                        foreach (var kvp in currentBalances)
-                        {
-                            _caches.UpdateBalanceIndex(kvp.Key, isV1: true, kvp.Value);
-                            _caches.V1BalancesByAccountAndToken.Add(currentBlock, kvp.Key, kvp.Value);
-                        }
-                    }
-                    currentBlock = blockNumber;
-                }
-
-                // Update sender balance
-                if (from != "0x0000000000000000000000000000000000000000")
-                {
-                    var fromKey = $"{from.ToLowerInvariant()}:{tokenKey}";
-                    // Initialize from cache if we haven't seen this key yet in this block range
-                    if (!currentBalances.ContainsKey(fromKey))
-                    {
-                        _caches.V1BalancesByAccountAndToken.TryGetValue(fromKey, out var existingBalance);
-                        currentBalances[fromKey] = existingBalance;
-                    }
-                    currentBalances[fromKey] -= value;
-                }
-
-                // Update receiver balance
-                if (to != "0x0000000000000000000000000000000000000000")
-                {
-                    var toKey = $"{to.ToLowerInvariant()}:{tokenKey}";
-                    // Initialize from cache if we haven't seen this key yet in this block range
-                    if (!currentBalances.ContainsKey(toKey))
-                    {
-                        _caches.V1BalancesByAccountAndToken.TryGetValue(toKey, out var existingBalance);
-                        currentBalances[toKey] = existingBalance;
-                    }
-                    currentBalances[toKey] += value;
-                }
-
-                transferCount++;
-            }
-        }
-
-        // Flush balances for the last block (index-before-balance ordering)
-        if (currentBlock != -1)
-        {
-            foreach (var kvp in currentBalances)
-            {
-                _caches.UpdateBalanceIndex(kvp.Key, isV1: true, kvp.Value);
-                _caches.V1BalancesByAccountAndToken.Add(currentBlock, kvp.Key, kvp.Value);
-            }
-        }
-
-        if (transferCount > 0)
-        {
-            _logger.LogDebug("Processed {Count} V1 transfer events", transferCount);
-        }
-    }
-
-    private async Task ProcessV2TransfersAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        // Process V2 TransferSingle and TransferBatch events together in block order
-        const string sql = @"
-            SELECT * FROM (
-                SELECT
-                    ""from"",
-                    ""to"",
-                    ""tokenAddress"",
-                    value,
-                    ""blockNumber"",
-                    ""transactionIndex"",
-                    ""logIndex"",
-                    ""timestamp"",
-                    'Single' as type
-                FROM ""CrcV2_TransferSingle""
-                WHERE ""blockNumber"" >= @fromBlock AND ""blockNumber"" <= @toBlock
-
-                UNION ALL
-
-                SELECT
-                    ""from"",
-                    ""to"",
-                    ""tokenAddress"",
-                    value,
-                    ""blockNumber"",
-                    ""transactionIndex"",
-                    ""logIndex"",
-                    ""timestamp"",
-                    'Batch' as type
-                FROM ""CrcV2_TransferBatch""
-                WHERE ""blockNumber"" >= @fromBlock AND ""blockNumber"" <= @toBlock
-            ) AS combined
-            ORDER BY ""blockNumber"", ""transactionIndex"", ""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var transferCount = 0;
-        var currentBalances = new Dictionary<string, decimal>();
-        long currentBlock = -1;
-
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var from = reader.GetString(0);
-                var to = reader.GetString(1);
-                var tokenAddress = reader.GetString(2).ToLowerInvariant();
-                var valueBig = reader.GetFieldValue<BigInteger>(3);
-                var blockNumber = reader.GetInt64(4);
-                var timestamp = reader.GetInt64(7);
-
-                // Convert from wei (18 decimals) to token units using CirclesConverter for proper precision
-                decimal amount;
-                try
-                {
-                    amount = CirclesConverter.AttoCirclesToCircles(valueBig);
-                }
-                catch (OverflowException)
-                {
-                    _logger.LogWarning("Skipping V2 transfer with value {Value} that would overflow decimal", valueBig);
-                    continue;
-                }
-
-                if (blockNumber != currentBlock)
-                {
-                    if (currentBlock != -1)
-                    {
-                        // Flush balances for the previous block.
-                        // Update secondary index BEFORE writing balance so concurrent reads
-                        // see the token in the index before the balance appears (worst case:
-                        // momentary 0-balance entry, not an invisible balance).
-                        foreach (var kvp in currentBalances)
-                        {
-                            _caches.UpdateBalanceIndex(kvp.Key, isV1: false, kvp.Value);
-                            _caches.V2BalancesByAccountAndToken.Add(currentBlock, kvp.Key, kvp.Value);
-                        }
-                    }
-                    currentBlock = blockNumber;
-                }
-
-                var fromLower = from.ToLowerInvariant();
-                var toLower = to.ToLowerInvariant();
-
-                // Update sender balance — token must be valid (account may be stopped but still holds tokens)
-                if (from != "0x0000000000000000000000000000000000000000"
-                    && CirclesInvariants.IsValidToken(tokenAddress, _registrations, _wrapperLookup))
-                {
-                    var fromKey = $"{fromLower}:{tokenAddress}";
-                    if (!currentBalances.ContainsKey(fromKey))
-                    {
-                        _caches.V2BalancesByAccountAndToken.TryGetValue(fromKey, out var existingBalance);
-                        currentBalances[fromKey] = existingBalance;
-                    }
-                    currentBalances[fromKey] -= amount;
-                    _caches.V2LastActivity.Add(blockNumber, fromKey, timestamp);
-                }
-
-                // Update receiver balance — token must be valid
-                if (to != "0x0000000000000000000000000000000000000000"
-                    && CirclesInvariants.IsValidToken(tokenAddress, _registrations, _wrapperLookup))
-                {
-                    var toKey = $"{toLower}:{tokenAddress}";
-                    if (!currentBalances.ContainsKey(toKey))
-                    {
-                        _caches.V2BalancesByAccountAndToken.TryGetValue(toKey, out var existingBalance);
-                        currentBalances[toKey] = existingBalance;
-                    }
-                    currentBalances[toKey] += amount;
-                    _caches.V2LastActivity.Add(blockNumber, toKey, timestamp);
-                }
-
-                transferCount++;
-            }
-        }
-
-        // Flush balances for the last block (index-before-balance ordering)
-        if (currentBlock != -1)
-        {
-            foreach (var kvp in currentBalances)
-            {
-                _caches.UpdateBalanceIndex(kvp.Key, isV1: false, kvp.Value);
-                _caches.V2BalancesByAccountAndToken.Add(currentBlock, kvp.Key, kvp.Value);
-            }
-        }
-
-        if (transferCount > 0)
-        {
-            _logger.LogDebug("Processed {Count} V2 transfer events", transferCount);
-        }
-    }
-
-    private async Task ProcessV2Erc20WrapperTransfersAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        // Process V2 ERC20 Wrapper Transfer events incrementally
-        const string sql = @"
-            SELECT ""from"", ""to"", ""tokenAddress"", amount, ""blockNumber"", ""timestamp""
-            FROM ""CrcV2_Erc20WrapperTransfer""
-            WHERE ""blockNumber"" >= @fromBlock AND ""blockNumber"" <= @toBlock
-            ORDER BY ""blockNumber"", ""transactionIndex"", ""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var transferCount = 0;
-        var currentBalances = new Dictionary<string, decimal>();
-        long currentBlock = -1;
-
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var from = reader.GetString(0);
-                var to = reader.GetString(1);
-                var tokenAddress = reader.GetString(2);
-                var amountBig = reader.GetFieldValue<BigInteger>(3);
-                var blockNumber = reader.GetInt64(4);
-                var timestamp = reader.GetInt64(5);
-
-                // Convert from wei (18 decimals) to token units using CirclesConverter for proper precision
-                decimal amount;
-                try
-                {
-                    amount = CirclesConverter.AttoCirclesToCircles(amountBig);
-                }
-                catch (OverflowException)
-                {
-                    _logger.LogWarning("Skipping ERC20 wrapper transfer with amount {Amount} that would overflow decimal", amountBig);
-                    continue;
-                }
-
-                var tokenKey = tokenAddress.ToLowerInvariant();
-
-                if (blockNumber != currentBlock)
-                {
-                    if (currentBlock != -1)
-                    {
-                        // Flush balances for the previous block (index-before-balance ordering)
-                        foreach (var kvp in currentBalances)
-                        {
-                            _caches.UpdateBalanceIndex(kvp.Key, isV1: false, kvp.Value);
-                            _caches.V2BalancesByAccountAndToken.Add(currentBlock, kvp.Key, kvp.Value);
-                        }
-                    }
-                    currentBlock = blockNumber;
-                }
-
-                var fromLower = from.ToLowerInvariant();
-                var toLower = to.ToLowerInvariant();
-
-                // Update sender balance — token must be valid (account may be stopped)
-                if (from != "0x0000000000000000000000000000000000000000"
-                    && CirclesInvariants.IsValidToken(tokenKey, _registrations, _wrapperLookup))
-                {
-                    var fromKey = $"{fromLower}:{tokenKey}";
-                    if (!currentBalances.ContainsKey(fromKey))
-                    {
-                        _caches.V2BalancesByAccountAndToken.TryGetValue(fromKey, out var existingBalance);
-                        currentBalances[fromKey] = existingBalance;
-                    }
-                    currentBalances[fromKey] -= amount;
-                    _caches.V2LastActivity.Add(blockNumber, fromKey, timestamp);
-                }
-
-                // Update receiver balance — token must be valid
-                if (to != "0x0000000000000000000000000000000000000000"
-                    && CirclesInvariants.IsValidToken(tokenKey, _registrations, _wrapperLookup))
-                {
-                    var toKey = $"{toLower}:{tokenKey}";
-                    if (!currentBalances.ContainsKey(toKey))
-                    {
-                        _caches.V2BalancesByAccountAndToken.TryGetValue(toKey, out var existingBalance);
-                        currentBalances[toKey] = existingBalance;
-                    }
-                    currentBalances[toKey] += amount;
-                    _caches.V2LastActivity.Add(blockNumber, toKey, timestamp);
-                }
-
-                transferCount++;
-            }
-        }
-
-        // Flush balances for the last block (index-before-balance ordering)
-        if (currentBlock != -1)
-        {
-            foreach (var kvp in currentBalances)
-            {
-                _caches.UpdateBalanceIndex(kvp.Key, isV1: false, kvp.Value);
-                _caches.V2BalancesByAccountAndToken.Add(currentBlock, kvp.Key, kvp.Value);
-            }
-        }
-
-        if (transferCount > 0)
-        {
-            _logger.LogDebug("Processed {Count} ERC20 wrapper transfer events", transferCount);
-        }
-    }
-
-    private async Task ProcessV1TrustAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        // Process V1 Trust events - update V1TrustRelations cache
-        const string sql = @"
-            SELECT t.""blockNumber"", t.""canSendTo"" as truster, t.""user"" as trustee, t.""limit""
-            FROM ""CrcV1_Trust"" t
-            WHERE t.""blockNumber"" >= @fromBlock AND t.""blockNumber"" <= @toBlock
-            ORDER BY t.""blockNumber"", t.""transactionIndex"", t.""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var count = 0;
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var blockNumber = reader.GetInt64(0);
-                var truster = reader.GetString(1);
-                var trustee = reader.GetString(2);
-                var limitBig = reader.GetFieldValue<BigInteger>(3);
-                long limitLong = limitBig > long.MaxValue ? long.MaxValue : (long)limitBig;
-
-                var key = $"{truster.ToLowerInvariant()}:{trustee.ToLowerInvariant()}";
-
-                // If limit is 0, remove the trust relation; otherwise add/update it
-                if (limitBig == 0)
-                {
-                    _caches.RemoveV1Trust(blockNumber, truster, trustee);
-                }
-                else
-                {
-                    // V1 trust stores the trust limit (0-100)
-                    _caches.UpsertV1Trust(blockNumber, truster, trustee, limitLong);
-                }
-
-                count++;
-            }
-        }
-
-        if (count > 0)
-        {
-            _logger.LogDebug("Processed {Count} V1 trust events", count);
-        }
-    }
-
-    private async Task ProcessV1UpdateMetadataDigestAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        // Process V1 UpdateMetadataDigest events - update V1AvatarToCidMap cache
-        const string sql = @"
-            SELECT m.""blockNumber"", m.avatar, m.""metadataDigest""
-            FROM ""CrcV1_UpdateMetadataDigest"" m
-            WHERE m.""blockNumber"" >= @fromBlock AND m.""blockNumber"" <= @toBlock
-            ORDER BY m.""blockNumber"", m.""transactionIndex"", m.""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var count = 0;
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var blockNumber = reader.GetInt64(0);
-                var avatar = reader.GetString(1);
-                var metadataDigest = (byte[])reader.GetValue(2);
-
-                var cid = CidHelper.MetadataDigestToCidV0(metadataDigest);
-                var key = avatar.ToLowerInvariant();
-
-                _caches.V1AvatarToCidMap.Add(blockNumber, key, cid);
-
-                count++;
-            }
-        }
-
-        if (count > 0)
-        {
-            _logger.LogDebug("Processed {Count} V1 metadata digest updates", count);
-        }
-    }
-
-    private async Task ProcessV2UpdateMetadataDigestAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        // Process V2 UpdateMetadataDigest events - update V2AvatarToCidMap cache
-        const string sql = @"
-            SELECT m.""blockNumber"", m.avatar, m.""metadataDigest""
-            FROM ""CrcV2_UpdateMetadataDigest"" m
-            WHERE m.""blockNumber"" >= @fromBlock AND m.""blockNumber"" <= @toBlock
-            ORDER BY m.""blockNumber"", m.""transactionIndex"", m.""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var count = 0;
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var blockNumber = reader.GetInt64(0);
-                var avatar = reader.GetString(1);
-                var metadataDigest = (byte[])reader.GetValue(2);
-
-                var cid = CidHelper.MetadataDigestToCidV0(metadataDigest);
-                var key = avatar.ToLowerInvariant();
-
-                _caches.V2AvatarToCidMap.Add(blockNumber, key, cid);
-
-                count++;
-            }
-        }
-
-        if (count > 0)
-        {
-            _logger.LogDebug("Processed {Count} V2 metadata digest updates", count);
-        }
-    }
-
-    private async Task ProcessV2RegisterShortNameAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        // Process V2 RegisterShortName events - update V2AvatarToShortNameMap cache
-        const string sql = @"
-            SELECT s.""blockNumber"", s.avatar, s.""shortName""
-            FROM ""CrcV2_RegisterShortName"" s
-            WHERE s.""blockNumber"" >= @fromBlock AND s.""blockNumber"" <= @toBlock
-            ORDER BY s.""blockNumber"", s.""transactionIndex"", s.""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var count = 0;
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var blockNumber = reader.GetInt64(0);
-                var avatar = reader.GetString(1);
-                // shortName is stored as numeric (BigInteger) in database
-                var shortNameNumeric = reader.GetFieldValue<BigInteger>(2);
-                // Convert to Base58Btc format (like "zAlice")
-                var shortNameBase58 = shortNameNumeric.ToBase58Btc();
-
-                var key = avatar.ToLowerInvariant();
-
-                _caches.V2AvatarToShortNameMap.Add(blockNumber, key, shortNameBase58);
-
-                count++;
-            }
-        }
-
-        if (count > 0)
-        {
-            _logger.LogDebug("Processed {Count} V2 short name registrations", count);
-        }
-    }
-
-    private async Task ProcessV2SetAdvancedUsageFlagAsync(NpgsqlConnection conn, long fromBlock, long toBlock, CancellationToken ct)
-    {
-        const string sql = @"
-            SELECT f.""blockNumber"", f.avatar, f.flag
-            FROM ""CrcV2_SetAdvancedUsageFlag"" f
-            WHERE f.""blockNumber"" >= @fromBlock AND f.""blockNumber"" <= @toBlock
-            ORDER BY f.""blockNumber"", f.""transactionIndex"", f.""logIndex""";
-
-        await using var cmd = new NpgsqlCommand(sql, conn);
-        cmd.Parameters.AddWithValue("fromBlock", fromBlock);
-        cmd.Parameters.AddWithValue("toBlock", toBlock);
-
-        var count = 0;
-        await using (var reader = await cmd.ExecuteReaderAsync(ct))
-        {
-            while (await reader.ReadAsync(ct))
-            {
-                var blockNumber = reader.GetInt64(0);
-                var avatar = reader.GetString(1).ToLowerInvariant();
-                var flag = (byte[])reader.GetValue(2);
-
-                // Registration check: only store flags for registered avatars
-                if (_registrations.IsRegistered(avatar))
-                {
-                    _caches.ConsentedFlowFlags.Add(blockNumber, avatar, flag);
-                    count++;
-                }
-            }
-        }
-
-        if (count > 0)
-        {
-            _logger.LogDebug("Processed {Count} V2 advanced usage flag updates", count);
-        }
     }
 }
